fixed: quickfix was modified refactored: inspection code, SmkRuleOrCheckpointArgsSection got new function
We create separate method in SmkRuleOrCheckpointArgsSectionImpl in order not to duplicate code, if there any other SmkRuleOrCheckpointArgsSection inheritor, which will need the same logic e.g. 'use' section. Other variant is to move multilineSectionDefinition() implementation into SmkRuleOrCheckpointArgsSection interface but I'm not sure if it is a good idea to mix definition and implementation.

Resolves: #148